Added null guards for public methods inside `ComponentDestructionService`.

+++ Runtime/DestructionService/ComponentDestructionService.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using System;
+using UnityEngine;
 
 public class ComponentDestructionService<T> : IDestructionService<T> where T : Component
 {
@@ -6,11 +7,17 @@ public class ComponentDestructionService<T> : IDestructionService<T> where T : C
 
 	public ComponentDestructionService(UnityObjectDestructionService<GameObject> unityObjectDestructionService)
 	{
+		if (unityObjectDestructionService == null)
+			throw new ArgumentNullException(nameof(unityObjectDestructionService));
+
 		this.unityObjectDestructionService = unityObjectDestructionService;
 	}
 
 	public void Destroy(T @object)
 	{
+		if (@object == null)
+			throw new ArgumentNullException(nameof(@object));
+
 		unityObjectDestructionService.Destroy(@object.gameObject);
 	}
 }
